requirements.txt
fastapi==0.115.5
pydantic==2.6.4
pydantic_core==2.16.3
uvicorn==0.32.1
google-generativeai==0.4.1